Se ajusta formulario para que logre guardar y se cambia tipo de dato en el modelo para el precio

+++ ApiProducto/Models/Producto.cs
@@ -13,7 +13,7 @@ namespace ApiProducto.Models
         [Required]
         public string Descripcion { get; set; }
         [Required]
-        public long Precio { get; set; }
+        public decimal Precio { get; set; }
         public DateTime FechaCreacion { get; set; }
         public bool Estado { get; set; }
     }
